APIO-128 Move ManagerCache to cache package
